app/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.23.2
openai==0.25.0
openpyxl==3.0.10
pandas==1.4.4
pandas-stubs==1.4.3.220829
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.2.1.0
typing_extensions==4.3.0
urllib3==1.26.12
